+++ frontend/src/components/control/add/index.js
@@ -1,0 +1,3 @@
+import PairedToogleButtons from './pairedToggleButtons'
+
+export default PairedToogleButtons
