+++ frontend/src/components/control/view/__test__/ControlView-test.js
@@ -1,4 +1,4 @@
-import { assert, React, ReactTestUtils } from '../../../../test/setup';
+import { assert, React, ReactTestUtils } from '../../../../../test/setup';
 import ControlView from '../ControlView';
 
 describe('ControlView', () => {
