+++ frontend/src/components/control/view/index.js
