+++ frontend/src/containers/control/add/index.js
@@ -20,10 +20,10 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { Grid, Row, Col, Button, FormControl, PageHeader } from 'react-bootstrap'
-import PairedToogleButtons from '../../components/controls/pairedToggleButtons'
-import Field from '../../components/formfields/musitfield'
-import { addControl } from '../../reducers/control'
-import Language from '../../components/language'
+import PairedToogleButtons from '../../../components/control/add'
+import Field from '../../../components/formfields/musitfield'
+import { addControl } from '../../../reducers/control/add'
+import Language from '../../../components/language'
 import DatePicker from 'react-bootstrap-date-picker'
 import moment from 'moment'
