+++ frontend/src/reducers/control/add/index.js
